requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
pymupdf
python-multipart
